[Stella][BugFix] Fix rss_rowid truncation in inline pk index compaction Link: https://code.alibaba-inc.com/soe/stella/codereview/27456550 * [Stella][BugFix] Fix rss_rowid truncation in inline pk index compaction
Two fixes to commit 48cb6f0518b ("[Stella][Enhancement] Optimize lake pk
table publish by inline major compacting sst files"):

1. Type truncation bug. PersistentIndexMemtable::max_rss_rowid() returns a
   uint64_t encoded as (rssid << 32) | rowid, but the new code in
   LakePersistentIndex::upsert() and the member _current_txn_rss_rowid_min
   were declared as uint32_t. The high 32 bits (rssid) were silently
   dropped, leaving only the rowid as the per-transaction lower bound.
   The check in try_inline_major_compaction()
       sst->max_rss_rowid() >= _current_txn_rss_rowid_min
   then trivially holds for every existing sstable, so on tablets with a
   pre-existing sstable backlog (>= 10 files) inline compaction degenerates
   into a full base + cumulative major compaction running synchronously on
   the publish thread. This adds large OSS round-trips to ordinary write
   publishes and can slow them by orders of magnitude. Fix by widening
   both the local variable and the member to uint64_t (and the reset in
   commit() to UINT64_MAX). Same fix as upstream commit 71955afd65a.

2. Flip lake_pk_index_publish_enable_inline_major_compaction default from
   "true" to "false". Even after the truncation fix, the optimization
   shifts heavy I/O onto the publish thread (one major_compact per
   memtable flush, inputs re-merged across rounds, no memory backpressure
   guard, races with the async pk-index compaction path). Keep it off by
   default until those risks are validated under representative
   workloads; the config is mutable so it can still be opted into per
   cluster at runtime.

+++ be/src/storage/lake/lake_persistent_index.cpp
@@ -308,7 +308,7 @@ Status LakePersistentIndex::upsert(size_t n, const Slice* keys, const IndexValue
     KeyIndexSet& key_indexes = not_founds;
     RETURN_IF_ERROR(get_from_sstables(n, keys, old_values, &key_indexes, -1));
     if (is_memtable_full()) {
-        uint32_t max_rss_rowid = _memtable->max_rss_rowid();
+        uint64_t max_rss_rowid = _memtable->max_rss_rowid();
         _current_txn_rss_rowid_min = std::min(_current_txn_rss_rowid_min, max_rss_rowid);
         RETURN_IF_ERROR(flush_memtable());
         if (config::lake_pk_index_publish_enable_inline_major_compaction) {
@@ -649,7 +649,7 @@ Status LakePersistentIndex::commit(MetaFileBuilder* builder) {
     _need_rebuild_file_cnt = need_rebuild_file_cnt(*builder->tablet_meta(), sstable_meta);
 
     // Reset transaction-level rssid tracking after commit
-    _current_txn_rss_rowid_min = UINT32_MAX;
+    _current_txn_rss_rowid_min = UINT64_MAX;
 
     return Status::OK();
 }
